+++ features/rum/src/appleMain/kotlin/com/datadog/kmp/rum/Rum.kt
@@ -6,10 +6,11 @@
 
 package com.datadog.kmp.rum
 
-import cocoapods.DatadogObjc.DDRUM
-import cocoapods.DatadogObjc.DDRUMConfiguration
-import cocoapods.DatadogObjc.DDRUMErrorSourceSource
-import cocoapods.DatadogObjc.DDRUMMonitor
+import cocoapods.DatadogRUM.DDRUM
+import cocoapods.DatadogRUM.DDRUMConfiguration
+import cocoapods.DatadogRUM.DDRUMErrorSourceSource
+import cocoapods.DatadogRUM.DDRUMMonitor
+import cocoapods.DatadogRUM._internal_sync_addError
 import com.datadog.kmp.internal.INCLUDE_BINARY_IMAGES
 import com.datadog.kmp.internal.InternalProxy
 import com.datadog.kmp.internal.RUM_ERROR_IS_CRASH
@@ -33,7 +34,7 @@ actual object Rum {
         if (InternalProxy.isCrashReportingEnabled) {
             addDatadogUnhandledExceptionHook {
                 DDRUMMonitor.shared()
-                    .addErrorWithError(
+                    ._internal_sync_addError(
                         createNSErrorFromThrowable(it),
                         DDRUMErrorSourceSource,
                         mutableMapOf<Any?, Any?>().apply {