+++ features/rum/src/commonMain/kotlin/com/datadog/kmp/rum/RumMonitor.kt
@@ -6,6 +6,8 @@
 
 package com.datadog.kmp.rum
 
+import com.datadog.kmp.rum.featureoperations.FailureReason
+
 /**
  *  A class enabling Datadog RUM features.
  *
@@ -225,6 +227,60 @@ interface RumMonitor {
      * @param key the attribute key (non null)
      */
     fun removeAttribute(key: String)
+
+    /**
+     * Starts the [name] feature operation.
+     *
+     * @param name the name of the operation.
+     * @param operationKey optional operation key. Allows to track multiple operations of the same [name].
+     * For example, multiple network requests (photo or file uploads) for the same URL.
+     * @param attributes additional custom attributes to attach to the feature operation.
+     */
+    @ExperimentalRumApi
+    fun startFeatureOperation(name: String, operationKey: String? = null, attributes: Map<String, Any?> = emptyMap())
+
+    /**
+     * Finishes the [name] feature operation with successful status.
+     *
+     * @param name the name of the operation.
+     * @param operationKey optional operation key identifying a specific operation
+     * instance from the list of feature operations of the same [name]. Should be provided if [operationKey] was
+     * provided during [startFeatureOperation] invocation.
+     * @param attributes additional custom attributes to attach to the feature operation. Can be
+     * used to add some result data produced as the result of the operation.
+     */
+    @ExperimentalRumApi
+    fun succeedFeatureOperation(name: String, operationKey: String? = null, attributes: Map<String, Any?> = emptyMap())
+
+    /**
+     * Finishes the [name] feature operation with failure status.
+     *
+     * @param name the name of the operation.
+     * @param operationKey optional operation key identifying a specific operation
+     * instance from the list of feature operations of the same [name]. Should be provided if [operationKey] was
+     * provided during [startFeatureOperation] invocation.
+     * @param failureReason the reason for the operation failure.
+     * @param attributes additional custom attributes to attach to the feature operation. Can be
+     * used to add some result data produced as the result of the operation.
+     */
+    @ExperimentalRumApi
+    fun failFeatureOperation(
+        name: String,
+        operationKey: String? = null,
+        failureReason: FailureReason,
+        attributes: Map<String, Any?> = emptyMap()
+    )
+
+    /**
+     * Adds view loading time to the active view based on the time elapsed since the view was started.
+     * The view loading time is automatically calculated as the difference between the current time
+     * and the start time of the view.
+     * This method should be called only once per view.
+     * If no view is started or active, this method does nothing.
+     * @param overwrite controls if the method overwrites the previously calculated view loading time.
+     */
+    @ExperimentalRumApi
+    fun addViewLoadingTime(overwrite: Boolean)
 
     /**
      * Stops the current session.