+++ features/session-replay/src/commonMain/kotlin/com/datadog/kmp/sessionreplay/configuration/internal/PlatformSessionReplayConfigurationBuilder.kt
@@ -7,13 +7,10 @@
 package com.datadog.kmp.sessionreplay.configuration.internal
 
 import com.datadog.kmp.sessionreplay.configuration.ImagePrivacy
-import com.datadog.kmp.sessionreplay.configuration.SessionReplayPrivacy
 import com.datadog.kmp.sessionreplay.configuration.TextAndInputPrivacy
 import com.datadog.kmp.sessionreplay.configuration.TouchPrivacy
 
 internal interface PlatformSessionReplayConfigurationBuilder<T : Any> {
-
-    fun setPrivacy(privacy: SessionReplayPrivacy)
 
     fun setImagePrivacy(privacy: ImagePrivacy)
 
@@ -22,6 +19,8 @@ internal interface PlatformSessionReplayConfigurationBuilder<T : Any> {
     fun setTextAndInputPrivacy(privacy: TextAndInputPrivacy)
 
     fun startRecordingImmediately(enabled: Boolean)
+
+    fun useCustomEndpoint(endpoint: String)
 
     fun build(): T
 }